fix: changed makefile to make cobertura work
