+++ cobertura_game.txt
@@ -1,0 +1,21 @@
+File 'src/game.c'
+Lines executed:83.33% of 12
+Branches executed:100.00% of 12
+Taken at least once:58.33% of 12
+Calls executed:60.00% of 5
+Creating 'game.c.gcov'
+
+File '/usr/include/stdlib.h'
+Lines executed:100.00% of 2
+Branches executed:100.00% of 2
+Taken at least once:50.00% of 2
+Calls executed:100.00% of 1
+Creating 'stdlib.h.gcov'
+
+File '/usr/include/x86_64-linux-gnu/bits/stdio2.h'
+Lines executed:0.00% of 2
+No branches
+Calls executed:0.00% of 1
+Creating 'stdio2.h.gcov'
+
+Lines executed:75.00% of 16
