+++ cobertura_io.txt
@@ -1,0 +1,15 @@
+File 'src/io.c'
+Lines executed:0.00% of 40
+Branches executed:0.00% of 40
+Taken at least once:0.00% of 40
+Calls executed:0.00% of 15
+Creating 'io.c.gcov'
+
+File '/usr/include/x86_64-linux-gnu/bits/stdio2.h'
+Lines executed:0.00% of 11
+Branches executed:0.00% of 20
+Taken at least once:0.00% of 20
+Calls executed:0.00% of 5
+Creating 'stdio2.h.gcov'
+
+Lines executed:0.00% of 51
