--- coverage-report/coverage.info
@@ -1,109 +1,0 @@
-TN:
-SF:/home/lucasfariapinto/Desktop/C/LI2/projeto/Shitori-LI2/src/board.c
-FN:15,mostrarTabuleiro
-FNDA:3,mostrarTabuleiro
-FN:21,carregarTabuleiro
-FNDA:2,carregarTabuleiro
-FN:44,gravarTabuleiro
-FNDA:1,gravarTabuleiro
-FN:58,modificarTabuleiro
-FNDA:2,modificarTabuleiro
-DA:15,3
-DA:16,12
-DA:17,9
-DA:19,3
-DA:21,2
-DA:22,2
-DA:23,2
-DA:24,2
-DA:25,2
-DA:27,2
-DA:28,0
-DA:29,0
-DA:32,7
-DA:33,6
-DA:34,1
-DA:35,1
-DA:39,1
-DA:40,1
-DA:41,1
-DA:44,1
-DA:45,1
-DA:46,1
-DA:47,1
-DA:48,1
-DA:50,1
-DA:51,3
-DA:52,2
-DA:54,1
-DA:55,1
-DA:58,2
-DA:59,2
-DA:60,2
-DA:62,2
-DA:63,0
-DA:64,0
-DA:67,2
-DA:68,2
-end_of_record
-TN:
-SF:/home/lucasfariapinto/Desktop/C/LI2/projeto/Shitori-LI2/tests/test_board.c
-FN:7,test_carregarTabuleiro_valido
-FNDA:1,test_carregarTabuleiro_valido
-FN:15,test_carregarTabuleiro_invalido
-FNDA:1,test_carregarTabuleiro_invalido
-FN:21,test_modificarTabuleiro_pintardebranco
-FNDA:1,test_modificarTabuleiro_pintardebranco
-FN:35,test_modificarTabuleiro_risca
-FNDA:1,test_modificarTabuleiro_risca
-FN:49,test_gravarTabuleiro
-FNDA:1,test_gravarTabuleiro
-FN:72,main
-FNDA:1,main
-DA:7,1
-DA:8,1
-DA:9,1
-DA:10,1
-DA:11,1
-DA:12,1
-DA:13,1
-DA:15,1
-DA:16,1
-DA:17,1
-DA:18,1
-DA:19,1
-DA:21,1
-DA:22,1
-DA:31,1
-DA:32,1
-DA:33,1
-DA:35,1
-DA:36,1
-DA:45,1
-DA:46,1
-DA:47,1
-DA:49,1
-DA:50,1
-DA:59,1
-DA:60,1
-DA:62,1
-DA:63,1
-DA:64,1
-DA:65,1
-DA:66,1
-DA:67,1
-DA:68,1
-DA:70,1
-DA:72,1
-DA:73,1
-DA:75,1
-DA:77,1
-DA:78,1
-DA:79,1
-DA:80,1
-DA:81,1
-DA:83,1
-DA:84,1
-DA:85,1
-DA:86,1
-end_of_record
