--- coverage-report/src/index.html
@@ -1,93 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-
-<html lang="en">
-
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-  <title>LCOV - coverage.info - src</title>
-  <link rel="stylesheet" type="text/css" href="../gcov.css">
-</head>
-
-<body>
-
-  <table width="100%" border=0 cellspacing=0 cellpadding=0>
-    <tr><td class="title">LCOV - code coverage report</td></tr>
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-
-    <tr>
-      <td width="100%">
-        <table cellpadding=1 border=0 width="100%">
-          <tr>
-            <td width="10%" class="headerItem">Current view:</td>
-            <td width="35%" class="headerValue"><a href="../index.html">top level</a> - src</td>
-            <td width="5%"></td>
-            <td width="15%"></td>
-            <td width="10%" class="headerCovTableHead">Hit</td>
-            <td width="10%" class="headerCovTableHead">Total</td>
-            <td width="15%" class="headerCovTableHead">Coverage</td>
-          </tr>
-          <tr>
-            <td class="headerItem">Test:</td>
-            <td class="headerValue">coverage.info</td>
-            <td></td>
-            <td class="headerItem">Lines:</td>
-            <td class="headerCovTableEntry">33</td>
-            <td class="headerCovTableEntry">37</td>
-            <td class="headerCovTableEntryMed">89.2 %</td>
-          </tr>
-          <tr>
-            <td class="headerItem">Date:</td>
-            <td class="headerValue">2025-04-16 12:58:34</td>
-            <td></td>
-            <td class="headerItem">Functions:</td>
-            <td class="headerCovTableEntry">4</td>
-            <td class="headerCovTableEntry">4</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
-          </tr>
-          <tr><td><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-        </table>
-      </td>
-    </tr>
-
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-  </table>
-
-  <center>
-  <table width="80%" cellpadding=1 cellspacing=1 border=0>
-
-    <tr>
-      <td width="50%"><br></td>
-      <td width="10%"></td>
-      <td width="10%"></td>
-      <td width="10%"></td>
-      <td width="10%"></td>
-      <td width="10%"></td>
-    </tr>
-
-    <tr>
-      <td class="tableHead">Filename <span class="tableHeadSort"><img src="../glass.png" width=10 height=14 alt="Sort by name" title="Sort by name" border=0></span></td>
-      <td class="tableHead" colspan=3>Line Coverage <span class="tableHeadSort"><a href="index-sort-l.html"><img src="../updown.png" width=10 height=14 alt="Sort by line coverage" title="Sort by line coverage" border=0></a></span></td>
-      <td class="tableHead" colspan=2>Functions <span class="tableHeadSort"><a href="index-sort-f.html"><img src="../updown.png" width=10 height=14 alt="Sort by function coverage" title="Sort by function coverage" border=0></a></span></td>
-    </tr>
-    <tr>
-      <td class="coverFile"><a href="board.c.gcov.html">board.c</a></td>
-      <td class="coverBar" align="center">
-        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="../amber.png" width=89 height=10 alt="89.2%"><img src="../snow.png" width=11 height=10 alt="89.2%"></td></tr></table>
-      </td>
-      <td class="coverPerMed">89.2&nbsp;%</td>
-      <td class="coverNumMed">33 / 37</td>
-      <td class="coverPerHi">100.0&nbsp;%</td>
-      <td class="coverNumHi">4 / 4</td>
-    </tr>
-  </table>
-  </center>
-  <br>
-
-  <table width="100%" border=0 cellspacing=0 cellpadding=0>
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php">LCOV version 1.14</a></td></tr>
-  </table>
-  <br>
-
-</body>
-</html>
