--- coverage-report/tests/test_board.c.func.html
@@ -1,96 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-
-<html lang="en">
-
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-  <title>LCOV - coverage.info - tests/test_board.c - functions</title>
-  <link rel="stylesheet" type="text/css" href="../gcov.css">
-</head>
-
-<body>
-
-  <table width="100%" border=0 cellspacing=0 cellpadding=0>
-    <tr><td class="title">LCOV - code coverage report</td></tr>
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-
-    <tr>
-      <td width="100%">
-        <table cellpadding=1 border=0 width="100%">
-          <tr>
-            <td width="10%" class="headerItem">Current view:</td>
-            <td width="35%" class="headerValue"><a href="../index.html">top level</a> - <a href="index.html">tests</a> - test_board.c<span style="font-size: 80%;"> (<a href="test_board.c.gcov.html">source</a> / functions)</span></td>
-            <td width="5%"></td>
-            <td width="15%"></td>
-            <td width="10%" class="headerCovTableHead">Hit</td>
-            <td width="10%" class="headerCovTableHead">Total</td>
-            <td width="15%" class="headerCovTableHead">Coverage</td>
-          </tr>
-          <tr>
-            <td class="headerItem">Test:</td>
-            <td class="headerValue">coverage.info</td>
-            <td></td>
-            <td class="headerItem">Lines:</td>
-            <td class="headerCovTableEntry">46</td>
-            <td class="headerCovTableEntry">46</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
-          </tr>
-          <tr>
-            <td class="headerItem">Date:</td>
-            <td class="headerValue">2025-04-16 12:58:34</td>
-            <td></td>
-            <td class="headerItem">Functions:</td>
-            <td class="headerCovTableEntry">6</td>
-            <td class="headerCovTableEntry">6</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
-          </tr>
-          <tr><td><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-        </table>
-      </td>
-    </tr>
-
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-  </table>
-
-  <center>
-  <table width="60%" cellpadding=1 cellspacing=1 border=0>
-    <tr><td><br></td></tr>
-    <tr>
-      <td width="80%" class="tableHead">Function Name <span class="tableHeadSort"><img src="../glass.png" width=10 height=14 alt="Sort by function name" title="Sort by function name" border=0></span></td>
-      <td width="20%" class="tableHead">Hit count <span class="tableHeadSort"><a href="test_board.c.func-sort-c.html"><img src="../updown.png" width=10 height=14 alt="Sort by hit count" title="Sort by hit count" border=0></a></span></td>
-    </tr>
-    <tr>
-              <td class="coverFn"><a href="test_board.c.gcov.html#70">main</a></td>
-              <td class="coverFnHi">1</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="test_board.c.gcov.html#13">test_carregarTabuleiro_invalido</a></td>
-              <td class="coverFnHi">1</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="test_board.c.gcov.html#5">test_carregarTabuleiro_valido</a></td>
-              <td class="coverFnHi">1</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="test_board.c.gcov.html#47">test_gravarTabuleiro</a></td>
-              <td class="coverFnHi">1</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="test_board.c.gcov.html#19">test_modificarTabuleiro_pintardebranco</a></td>
-              <td class="coverFnHi">1</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="test_board.c.gcov.html#33">test_modificarTabuleiro_risca</a></td>
-              <td class="coverFnHi">1</td>
-            </tr>
-  </table>
-  <br>
-  </center>
-  <table width="100%" border=0 cellspacing=0 cellpadding=0>
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
-  </table>
-  <br>
-
-</body>
-</html>
