+++ game.c.gcov
@@ -1,0 +1,51 @@
+        -:    0:Source:src/game.c
+        -:    0:Graph:test_game-game.gcno
+        -:    0:Data:test_game-game.gcda
+        -:    0:Runs:1
+        -:    1:#include <stdio.h>
+        -:    2:#include <stdlib.h>
+        -:    3:#include <ctype.h>
+        -:    4:#include "game.h"
+        -:    5:
+function modificarTabuleiro called 2 returned 100% blocks executed 86%
+        2:    6:void modificarTabuleiro(Tabuleiro *tab, Historico *hist, char cmd, const char *coord)
+        -:    7:{
+        2:    8:    int linha = atoi(&coord[1]) - 1;
+call    0 returned 2
+        2:    9:    int coluna = tolower(coord[0]) - 'a';
+        -:   10:
+        2:   11:    if (linha < 0 || linha >= tab->linhas || coluna < 0 || coluna >= tab->colunas)
+branch  0 taken 2 (fallthrough)
+branch  1 taken 0
+branch  2 taken 2 (fallthrough)
+branch  3 taken 0
+branch  4 taken 2 (fallthrough)
+branch  5 taken 0
+branch  6 taken 0 (fallthrough)
+branch  7 taken 2
+        -:   12:    {
+    #####:   13:        printf("Coordenada inválida.\n");
+call    0 never executed
+    #####:   14:        return;
+call    0 never executed
+        -:   15:    }
+        -:   16:
+        2:   17:    guardar_estado(hist, tab);
+call    0 returned 2
+        -:   18:
+        2:   19:    if (cmd == 'b')
+branch  0 taken 1 (fallthrough)
+branch  1 taken 1
+        -:   20:    {
+        1:   21:        tab->grelha[linha][coluna] = toupper(tab->grelha[linha][coluna]);
+        -:   22:    }
+        1:   23:    else if (cmd == 'r')
+branch  0 taken 1 (fallthrough)
+branch  1 taken 0
+        -:   24:    {
+        1:   25:        tab->grelha[linha][coluna] = '#';
+        -:   26:    }
+        -:   27:
+        2:   28:    mostrarTabuleiro(tab);
+call    0 returned 2
+        -:   29:}
